12.6.1-10 baseline
Former-commit-id: 5e2f5e9bc2b0559c3101cc0c3af708858e3b95d1

+++ cave/com.raytheon.uf.viz.d2d.nsharp/src/com/raytheon/uf/viz/d2d/nsharp/rsc/D2DNSharpResourceData.java
@@ -178,8 +178,8 @@ public abstract class D2DNSharpResourceData extends
             stnInfo.setRangestarttime(fcstTime);
         }
         if (coordinate != null) {
-            stnInfo.setLongitude(coordinate.x);
-            stnInfo.setLatitude(coordinate.y);
+            stnInfo.setLongitude((float) coordinate.x);
+            stnInfo.setLatitude((float) coordinate.y);
         }
         if (pointName != null) {
             stnInfo.setStnDisplayInfo(pointName + " "
